Improvised security features

+++ app/controllers/courses_controller.rb
@@ -1,6 +1,6 @@
 class CoursesController < ApplicationController
 
-    
+    before_action :authenticate_user!
     before_action :check_user
     
     def check_user
@@ -8,10 +8,10 @@ class CoursesController < ApplicationController
             flash[:notice]='Restricted Access'
             redirect_to root_path
         end
-    end
+    end 
 
     def index
-        @course=Course.where(college_id: params[:college_id])
+        @courses=Course.where(college_id: current_user.college.id)
     end
     
     def new
@@ -20,7 +20,7 @@ class CoursesController < ApplicationController
     end
     def create
         @course=Course.new(course_params)
-        @course.college_id=params[:college_id]
+        @course.college_id=current_user.college.id
         if @course.save
             flash[:notice]='Saved Successfully!'
             redirect_to root_path
@@ -35,31 +35,61 @@ class CoursesController < ApplicationController
     end
 
     def edit
-        @course=Course.where(params[:id])
-        @college = College.find(params[:college_id])
+        @course=Course.find(params[:id])
+        if @course
+            if @course.college_id==current_user.college.id
+                @college = College.find(params[:college_id])
+                render :edit
+            else
+                flash[:notice]='Restricted Access'
+                redirect_to root_path
+            end
+        else
+            flash[:notice]='Course not found'
+        end
     end
     def update
         @course=Course.find(params[:id])
-        @course.college_id=params[:college_id]
-        if @course.update(course_params)
-            flash[:notice]='Updated Successfully!'
-            redirect_to root_path
-        else
-            if @course.errors.any? 
-                @course.errors.full_messages.each do |message|
-                    flash[:notice]=message
+        if @course
+            if @course.college.id==current_user.college.id
+                
+                if @course.update(course_params)
+                    flash[:notice]='Updated Successfully!'
+                    redirect_to root_path
+                else
+                    if @course.errors.any? 
+                        @course.errors.full_messages.each do |message|
+                            flash[:notice]=message
+                        end
+                    end
+                    render :new, status: :unprocessable_entity
                 end
+            else
+                flash[:notice]='Restricted Access'
+                redirect_to root_path
             end
-            render :new, status: :unprocessable_entity
+        else
+            flash[:notice]='Not found'
         end
     end
 
     def destroy
         @course=Course.find(params[:id])
-        if @course.destroy
-            flash[:notice]='Deleted Successfully'
+        if @course
+            if @course.college.id==current_user.college.id
+                if @course.destroy
+                    flash[:notice]='Deleted Successfully'
+                    redirect_to root_path
+                else
+                    flash[:notice]='Please try again'
+                    redirect_to root_path
+                end
+            else
+                flash[:notice]='Restricted Access'
+                redirect_to root_path
+            end
         else
-            flash[:notice]='Please try again'
+            flash[:notice]='Course not found'
         end
 
     end
